% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YFinder.R
\name{call_ancestor}
\alias{call_ancestor}
\title{Call Ancestor}
\usage{
call_ancestor(hg, jsonnest, yfullsnp)
}
\arguments{
\item{hg}{The subclade eg R-Z2124}

\item{jsonnest}{the loaded JSON yfull file (use fromJSON() in package 
'jsonlite')}

\item{yfullsnp}{additional yfull snp list file provided in/data}
}
\value{
ancestor of the subclade
}
\description{
Find the immediate ancestor of a subclade as per provided yfull json tree
}
\examples{
anc <- call_ancestor("R-Z2124",loadedjsonobject,loadedyfullsnpfile);
}
