% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YFinder.R
\name{call_children}
\alias{call_children}
\title{Call Children}
\usage{
call_children(hg, jsonnest)
}
\arguments{
\item{hg}{The subclade eg R-Z2124}

\item{jsonnest}{the loaded JSON yfull file (use fromJSON() in package 
'jsonlite')}
}
\value{
ancestor of the subclade
}
\description{
Find the immediate children of a subclade as per provided yfull json tree
}
\examples{
anc <- call_children("R-Z2124",loadedjsonobject);
}
