% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YFinder.R
\name{depthfinder}
\alias{depthfinder}
\title{Depth Finder}
\usage{
depthfinder(hg, jsonnest)
}
\arguments{
\item{hg}{The subclade eg R-Z2124}

\item{jsonnest}{the loaded JSON yfull file (use jsonlite::fromJSON() in package 
'jsonlite')}
}
\value{
The depth in the YFull v11.02 json nested list
}
\description{
Find depth of a subclade on the YFull json tree (not actual depth, but depth
 in the json which is always more)
}
\examples{
depth <- depthfinder("R-Z2124",loadedjsonobject);
}
