% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YFinder.R
\name{YFinder}
\alias{YFinder}
\title{YFinder - Get haplogroups from VCF File}
\usage{
YFinder(
  vcffile,
  yfullpositionsfile,
  indivlistfile,
  yfulljsonfile,
  transversionsonly = F,
  depth = 1,
  callquality = 0.9,
  numsnptotal = 2,
  viewrealtime = FALSE
)
}
\arguments{
\item{vcffile}{The location of the vcf file}

\item{yfullpositionsfile}{The location of the yfullpositions file (
provided in /data folder)}

\item{indivlistfile}{the location of the file with individual sample ids 
to be tested. 1 sampleid per line. Last line should be empty. Ensure that
each id matches exactly with the sampleid column names in the vcf file.}

\item{yfulljsonfile}{location of the yfulljsonfile (provided in /data)}

\item{transversionsonly}{TRUE or FALSE. TRUE ignore C/T and G/A positions
Use TRUE for ancient samples to avoid false calls}

\item{depth}{number of ancestors to check in case of a single +ve SNP. 
default 2}

\item{callquality}{default 0.9. Percentage of Derived calls per subclade for 
assigning derived status to subclade.}

\item{numsnptotal}{Min number of SNPs to decide +ve subclade 
(without checking for ancestors). Default 2. Avoid 1.}

\item{viewrealtime}{Updates and shows the output table after every individual 
sample is processed for realtime view. Deafult is FALSE.}
}
\value{
Data frame object with list of assigned haplogroups and 
a call table to crossverify manually
}
\description{
Get list of all Y Hg calls based on YFull from VCF file of Y chr 
with multiple samples
}
